Results Section Tailwind Update
Converted results section to tailwind.
Modified the section containers for results section as it was in the calculator's section tag.  Added id class to results section to toggle invisibility upon button press.  Minor modifications to results section formatting.  Future formatting will occur for results section in separate commit.  The purpose of this commit is to ensure existing content is converted from CSS to Tailwind.